pyproject.toml
[project]
name = "ai-agents-mcp-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.50.0",
    "mcp>=1.12.3",
    "python-dotenv>=1.1.0",
]
